# Conflicts:
#	crates/backend/access/pgrcolumnar/src/loadsort.rs
#	crates/backend/storage/file/vfs/src/posix.rs
#	notes/LANES.md
#	scripts/lint-determinism.allow

+++ crates/backend/partitioning/partbounds/src/merge_tests.rs
@@ -12,7 +12,7 @@ fn static_mcx() -> Mcx<'static> {
 }
 
 fn img(v: i64) -> DatumImage<'static> {
-    DatumImage::ByVal(v as usize)
+    DatumImage::ByVal(v as u64)
 }
 
 fn img_val(d: &DatumImage<'static>) -> i64 {
